chore: rename BaseNewLayout to BaseLayout
